Upgrade to Astro 5 and convert to new Content Layer API
