+++ package.json
@@ -9,16 +9,17 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/node": "^8.3.4",
-    "@astrojs/starlight": "^0.29.0",
-    "@interledger/docs-design-system": "^0.5.2",
+    "@astrojs/mdx": "^4.0.2",
+    "@astrojs/node": "^9.0.0",
+    "@astrojs/starlight": "^0.30.1",
+    "@interledger/docs-design-system": "^0.5.5",
     "@types/showdown": "^2.0.6",
-    "astro": "^4.16.11",
+    "astro": "^5.0.5",
     "html-to-text": "^9.0.5",
     "markdown-it": "^14.1.0",
     "node-html-parser": "^6.1.13",
     "sharp": "^0.33.5",
     "showdown": "^2.1.0",
-    "starlight-links-validator": "^0.13.2"
+    "starlight-links-validator": "^0.13.4"
   }
 }
