+++ src/components/blog/LargeImg.astro
