+++ src/content.config.ts
@@ -1,0 +1,24 @@
+import { z, defineCollection } from "astro:content";
+import { glob } from "astro/loaders";
+import { docsLoader, i18nLoader } from "@astrojs/starlight/loaders";
+import { docsSchema, i18nSchema } from "@astrojs/starlight/schema";
+
+const blogCollection = defineCollection({
+  loader: glob({ pattern: "**/[^_]*.{md,mdx}", base: "./src/content/blog" }),
+  schema: z.object({
+    title: z.string(),
+    description: z.string(),
+    slug: z.string(),
+    date: z.date(),
+    image: z.string().optional(),
+    tags: z.array(z.string()),
+    authors: z.array(z.string()),
+    author_urls: z.array(z.string()),
+  }),
+});
+
+export const collections = {
+  docs: defineCollection({ loader: docsLoader(), schema: docsSchema() }),
+  i18n: defineCollection({ loader: i18nLoader(), schema: i18nSchema() }),
+  blog: blogCollection,
+};
