+++ src/content/blog/2018-10-03-interledger-how-to-interconnect-all-blockchains-and-value-networks.md
@@ -1,11 +1,12 @@
 ---
-layout: ../../layouts/BlogLayout.astro
 title: "Interledger: How to Interconnect All Blockchains and Value Networks"
 description: "Interledger was born out of a project to build a blockchain-agnostic smart contracts platform. A key challenge was neutrality: how could a decentralized app buy resources like storage and computing, without being tied to a specific blockchain?"
 date: 2018-10-03
 slug: interledger-how-to-interconnect-all-blockchains-and-value-networks
-authors: [Evan Schwartz]
-author_urls: [https://www.linkedin.com/in/evanmarkschwartz/]
+authors:
+  - Evan Schwartz
+author_urls:
+  - https://www.linkedin.com/in/evanmarkschwartz/
 external_url: https://medium.com/xpring/interledger-how-to-interconnect-all-blockchains-and-value-networks-74f432e64543
 tags:
   - Interledger
