+++ src/content/blog/2024-07-30-open-payments-cinderella-story.mdx
@@ -1,18 +1,19 @@
 ---
-layout: ../../layouts/BlogLayout.astro
 title: "Open Payments: The Cinderella Story of Finding a Fitting Authorization Method"
 description: A breakdown of the unique needs that an authorization method for Open Payments needs to be able to fulfill.
 date: 2024-07-30
 slug: open-payments-cinderella-story
-authors: [Nathan Lie]
-author_urls: [https://www.linkedin.com/in/nathan-lie-138a73121]
+authors:
+  - Nathan Lie
+author_urls:
+  - https://www.linkedin.com/in/nathan-lie-138a73121
 tags:
   - Interledger
   - Open Payments
   - GNAP
 ---
 
-import LargeImg from "/src/components/pages/LargeImg.astro";
+import LargeImg from "/src/components/blog/LargeImg.astro";
 
 ## The Internet Runs on OAuth 2.0
