+++ src/content/blog/2024-08-13-interledger-universe.mdx
@@ -1,12 +1,13 @@
 ---
-layout: ../../layouts/BlogLayout.astro
 title: "The Interledger Universe"
 description: "Or: “What the heck are all those products and protocols?”"
 ogImageUrl: /developers/img/blog/2024-08-13/og-image.png
 date: 2024-08-13
 slug: interledger-universe
-authors: [Sabine Schaller]
-author_urls: [https://www.linkedin.com/in/sabineschaller]
+authors:
+  - Sabine Schaller
+author_urls:
+  - https://www.linkedin.com/in/sabineschaller
 tags:
   - Interledger
   - Interledger Protocol
