+++ src/content/blog/2024-10-11-where-did-rafiki-money-go.mdx
@@ -1,11 +1,12 @@
 ---
-layout: ../../layouts/BlogLayout.astro
 title: "Where did rafiki.money go?"
 description: "Or “The need for rebranding when something confuses people.”"
 date: 2024-10-11
 slug: where-did-rafiki-money-go
-authors: [Timea Nagy]
-author_urls: [https://www.linkedin.com/in/nagy-timea-35483024]
+authors:
+  - Timea Nagy
+author_urls:
+  - https://www.linkedin.com/in/nagy-timea-35483024
 tags:
   - Test Network
 ---
@@ -44,13 +45,7 @@ Upon finalizing the visuals with our in-house illustrator, [Madalina Tantareanu]
 
 We were also able to account for designing the interactions, like hover states, which commonly get missed out if a static design file is used.
 
-<video
-  src="/developers/img/blog/2024-10-11/hover-effect.mp4"
-  controls
-  autoplay
-  loop
-  muted
-></video>
+<video src="/developers/img/blog/2024-10-11/hover-effect.mp4" controls autoplay loop muted></video>
 
 ## Challenges we faced, interesting stuff we used
