--- src/content/config.ts
@@ -1,19 +1,0 @@
-import { z, defineCollection } from "astro:content";
-import { docsSchema, i18nSchema } from "@astrojs/starlight/schema";
-
-const blogCollection = defineCollection({
-  type: "content",
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    date: z.date(),
-    image: z.string().optional(),
-    tags: z.array(z.string()),
-  }),
-});
-
-export const collections = {
-  docs: defineCollection({ schema: docsSchema() }),
-  i18n: defineCollection({ type: "data", schema: i18nSchema() }),
-  blog: blogCollection,
-};
