+++ src/layouts/BlogLayout.astro
@@ -209,4 +209,8 @@ article :global(code) {
 	padding: 0.125rem 0.375rem;
 	overflow-wrap: anywhere;
 }
+
+article :global(ul li) {
+	margin-block-end: var(--space-2xs);
+}
 </style>
