+++ src/pages/blog/[...id].astro
@@ -1,0 +1,21 @@
+---
+import { getCollection, render } from 'astro:content';
+import BlogLayout from '../../layouts/BlogLayout.astro';
+import CommunityLinks from "../../components/blog/CommunityLinks.astro";
+
+export async function getStaticPaths() {
+  const posts = await getCollection('blog');
+  return posts.map(post => ({
+    params: { id: post.id },
+    props: { post },
+  }));
+}
+
+const { post } = Astro.props;
+const { Content } = await render(post);
+---
+<BlogLayout frontmatter={post.data}>
+  <Content />
+  <CommunityLinks />
+</BlogLayout>
+
