--- src/pages/blog/[...slug].astro
@@ -1,14 +1,0 @@
----
-import { getCollection } from 'astro:content';
-
-export async function getStaticPaths() {
-  const blogEntries = await getCollection('blog');
-  return blogEntries.map(entry => ({
-    params: { slug: entry.slug }, props: { entry },
-  }));
-}
-
-const { entry } = Astro.props;
-const { Content } = await entry.render();
----
-<Content />
